LCAM-471-Calculate Evidence Fee

+++ crime-evidence/src/main/java/uk/gov/justice/laa/crime/evidence/dto/EvidenceFeeRulesDTO.java
@@ -1,6 +1,9 @@
 package uk.gov.justice.laa.crime.evidence.dto;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder
@@ -10,7 +13,7 @@ public class EvidenceFeeRulesDTO {
     private String emstCode;
     private String allIncomeEvidenceReceived;
     private String allCapitalEvidenceReceived;
-    private Integer capitalEvidenceItemsLower;
-    private Integer capitalEvidenceItemsUpper;
+    private Long capitalEvidenceItemsLower;
+    private Long capitalEvidenceItemsUpper;
 
 }
